Changed format visually and functionally
Visually instead of having 5 options that were buttons at the top of the screen, the buttoons were moved to the left side of the grid
leaving the title at the top. Functionally there are just 3 buttons now where one of them is for choosing the size by prompting the
user to type a number after clicking the button. Another button to reset the color and a another to choose the color to use.